+++ pkg/analyzer/lib/src/dart/resolver/scope_context.dart
@@ -247,9 +247,7 @@ class ScopeContext {
     if (functionTypedSuffix == null) {
       node.type?.accept(visitor);
     } else {
-      var element = node.declaredFragment!.element;
-      // ignore: deprecated_member_use_from_same_package
-      withTypeParameterScope(element.typeParameters, () {
+      withTypeParameterList(functionTypedSuffix.typeParameters, () {
         node.type?.accept(visitor);
         functionTypedSuffix.typeParameters?.accept(visitor);
         functionTypedSuffix.formalParameters.accept(visitor);