+++ pkg/analyzer_testing/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.3-dev
+
+- Require version `^14.0.0-0` of the `analyzer` package.
+
 ## 0.3.2
 
 - Require version `13.3.0` of the `analyzer` package.